pyproject.toml
[tool.poetry]
name = "fastapi-assignment-1"
version = "0.1.0"
description = ""
authors = ["Minkyu Lee <whizkyu@snu.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.36"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
pymysql = "^1.1.1"
cryptography = "^44.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
